chore(search): remove next folder
