+++ packages/react/src/components/Search/Search.js
@@ -5,283 +5,251 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import PropTypes from 'prop-types';
-import React, { Component } from 'react';
-import classNames from 'classnames';
 import { Search as SearchIcon, Close } from '@carbon/icons-react';
-import { composeEventHandlers } from '../../tools/events';
+import cx from 'classnames';
+import PropTypes from 'prop-types';
+import React, { useRef, useState } from 'react';
+import { focus } from '../../internal/focus';
 import { keys, match } from '../../internal/keyboard';
+import { useId } from '../../internal/useId';
+import { usePrefix } from '../../internal/usePrefix';
+import { composeEventHandlers } from '../../tools/events';
+import { useMergedRefs } from '../../internal/useMergedRefs';
 import deprecate from '../../prop-types/deprecate';
-import { FeatureFlagContext } from '../FeatureFlags';
-import { PrefixContext } from '../../internal/usePrefix';
 
-export default class Search extends Component {
-  static propTypes = {
-    /**
-     * Specify an optional className to be applied to the container node
-     */
-    className: PropTypes.string,
+const Search = React.forwardRef(function Search(
+  {
+    autoComplete = 'off',
+    className,
+    closeButtonLabelText = 'Clear search input',
+    defaultValue,
+    disabled,
+    id,
+    labelText,
+    light,
+    onChange = () => {},
+    onClear = () => {},
+    onKeyDown,
+    onExpand,
+    placeholder = '',
+    renderIcon,
+    role = 'searchbox',
+    size = 'md',
+    type = 'text',
+    value,
+    ...rest
+  },
+  forwardRef
+) {
+  const prefix = usePrefix();
+  const inputRef = useRef(null);
+  const ref = useMergedRefs([forwardRef, inputRef]);
+  const inputId = useId('search-input');
+  const uniqueId = id || inputId;
+  const searchId = `${uniqueId}-search`;
+  const [hasContent, setHasContent] = useState(value || defaultValue || false);
+  const [prevValue, setPrevValue] = useState(value);
+  const searchClasses = cx({
+    [`${prefix}--search`]: true,
+    [`${prefix}--search--sm`]: size === 'sm',
+    [`${prefix}--search--md`]: size === 'md',
+    [`${prefix}--search--lg`]: size === 'lg',
+    [`${prefix}--search--light`]: light,
+    [`${prefix}--search--disabled`]: disabled,
+    [className]: className,
+  });
+  const clearClasses = cx({
+    [`${prefix}--search-close`]: true,
+    [`${prefix}--search-close--hidden`]: !hasContent,
+  });
 
-    /**
-     * Specify a label to be read by screen readers on the "close" button
-     */
-    closeButtonLabelText: PropTypes.string,
-
-    /**
-     * Optionally provide the default value of the `<input>`
-     */
-    defaultValue: PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
-
-    /**
-     * Specify whether the `<input>` should be disabled
-     */
-    disabled: PropTypes.bool,
-
-    /**
-     * Specify a custom `id` for the input
-     */
-    id: PropTypes.string,
-
-    /**
-     * Provide the label text for the Search icon
-     */
-    labelText: PropTypes.node.isRequired,
-
-    /**
-     * Specify light version or default version of this control
-     */
-    light: PropTypes.bool,
-
-    /**
-     * Optional callback called when the search value changes.
-     */
-    onChange: PropTypes.func,
-
-    /**
-     * Optional callback called when the search value is cleared.
-     */
-    onClear: PropTypes.func,
-
-    /**
-     * Provide a handler that is invoked on the key down event for the input
-     */
-    onKeyDown: PropTypes.func,
-
-    /**
-     * Deprecated in favor of `placeholder`
-     */
-    placeHolderText: deprecate(
-      PropTypes.string,
-      `\nThe prop \`placeHolderText\` for Search has been deprecated in favor of \`placeholder\`. Please use \`placeholder\` instead.`
-    ),
-
-    /**
-     * Provide an optional placeholder text for the Search.
-     * Note: if the label and placeholder differ,
-     * VoiceOver on Mac will read both
-     */
-    placeholder: PropTypes.string,
-
-    /**
-     * Rendered icon for the Search.
-     * Can be a React component class
-     */
-    renderIcon: PropTypes.oneOfType([PropTypes.func, PropTypes.object]),
-
-    /**
-     * Specify the search size
-     */
-    size: PropTypes.oneOf(['sm', 'md', 'lg', 'xl']),
-
-    /**
-     * Specify whether the Search should be a small variant
-     */
-
-    /**
-     * Specify whether the load was successful
-     */
-    small: deprecate(
-      PropTypes.bool,
-      `\nThe prop \`small\` for Search has been deprecated in favor of \`size\`. Please use \`size="sm"\` instead.`
-    ),
-
-    /**
-     * Optional prop to specify the type of the `<input>`
-     */
-    type: PropTypes.string,
-
-    /**
-     * Specify the value of the `<input>`
-     */
-    value: PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
-  };
-
-  static contextType = FeatureFlagContext;
-
-  static defaultProps = {
-    type: 'text',
-    placeholder: '',
-    closeButtonLabelText: 'Clear search input',
-    onChange: () => {},
-    onClear: () => {},
-  };
-
-  state = {
-    hasContent: this.props.value || this.props.defaultValue || false,
-    prevValue: this.props.value,
-  };
-
-  static getDerivedStateFromProps({ value }, state) {
-    const { prevValue } = state;
-    return prevValue === value
-      ? null
-      : {
-          hasContent: !!value,
-          prevValue: value,
-        };
+  if (value !== prevValue) {
+    setHasContent(!!value);
+    setPrevValue(value);
   }
 
-  clearInput = (evt) => {
-    if (!this.props.value) {
-      this.input.value = '';
-      this.props.onChange(evt);
-    } else {
-      const clearedEvt = Object.assign({}, evt.target, {
-        target: {
-          value: '',
-        },
-      });
-      this.props.onChange(clearedEvt);
+  function clearInput() {
+    if (!value) {
+      inputRef.current.value = '';
     }
 
-    this.props.onClear();
+    const inputTarget = Object.assign({}, inputRef.current, { value: '' });
+    const clearedEvt = { target: inputTarget, type: 'change' };
 
-    this.setState({ hasContent: false }, () => this.input.focus());
-  };
+    onChange(clearedEvt);
+    onClear();
+    setHasContent(false);
+    focus(inputRef);
+  }
 
-  handleChange = (evt) => {
-    this.setState({
-      hasContent: evt.target.value !== '',
+  function handleChange(event) {
+    setHasContent(event.target.value !== '');
+  }
+
+  function handleKeyDown(event) {
+    if (match(event, keys.Escape)) {
+      event.stopPropagation();
+      clearInput(event);
+    }
+  }
+
+  return (
+    <div role="search" aria-labelledby={searchId} className={searchClasses}>
+      {/* the magnifier is used in ExpandableSearch as a click target to expand,
+      however, it does not need a keyboard event bc the input element gets focus on keyboard nav and expands that way*/}
+      {/* eslint-disable-next-line jsx-a11y/click-events-have-key-events */}
+      <div
+        role={onExpand ? 'button' : null}
+        className={`${prefix}--search-magnifier`}
+        onClick={onExpand}>
+        <CustomSearchIcon icon={renderIcon} />
+      </div>
+      <label id={searchId} htmlFor={uniqueId} className={`${prefix}--label`}>
+        {labelText}
+      </label>
+      <input
+        {...rest}
+        autoComplete={autoComplete}
+        className={`${prefix}--search-input`}
+        defaultValue={defaultValue}
+        disabled={disabled}
+        role={role}
+        ref={ref}
+        id={uniqueId}
+        onChange={composeEventHandlers([onChange, handleChange])}
+        onKeyDown={composeEventHandlers([onKeyDown, handleKeyDown])}
+        placeholder={placeholder}
+        type={type}
+        value={value}
+      />
+      <button
+        aria-label={closeButtonLabelText}
+        className={clearClasses}
+        disabled={disabled}
+        onClick={clearInput}
+        title={closeButtonLabelText}
+        type="button">
+        <Close />
+      </button>
+    </div>
+  );
+});
+
+Search.propTypes = {
+  /**
+   * Specify an optional value for the `autocomplete` property on the underlying
+   * `<input>`, defaults to "off"
+   */
+  autoComplete: PropTypes.string,
+
+  /**
+   * Specify an optional className to be applied to the container node
+   */
+  className: PropTypes.string,
+
+  /**
+   * Specify a label to be read by screen readers on the "close" button
+   */
+  closeButtonLabelText: PropTypes.string,
+
+  /**
+   * Optionally provide the default value of the `<input>`
+   */
+  defaultValue: PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
+
+  /**
+   * Specify whether the `<input>` should be disabled
+   */
+  disabled: PropTypes.bool,
+
+  /**
+   * Specify a custom `id` for the input
+   */
+  id: PropTypes.string,
+
+  /**
+   * Provide the label text for the Search icon
+   */
+  labelText: PropTypes.node.isRequired,
+
+  /**
+   * Specify light version or default version of this control
+   */
+  light: deprecate(
+    PropTypes.bool,
+    'The `light` prop for `Search` is no longer needed and has ' +
+      'been deprecated in v11 in favor of the new `Layer` component. It will be moved in the next major release.'
+  ),
+
+  /**
+   * Optional callback called when the search value changes.
+   */
+  onChange: PropTypes.func,
+
+  /**
+   * Optional callback called when the search value is cleared.
+   */
+  onClear: PropTypes.func,
+
+  /**
+   * Optional callback called when the magnifier icon is clicked in ExpandableSearch.
+   */
+  onExpand: PropTypes.func,
+
+  /**
+   * Provide a handler that is invoked on the key down event for the input
+   */
+  onKeyDown: PropTypes.func,
+
+  /**
+   * Provide an optional placeholder text for the Search.
+   * Note: if the label and placeholder differ,
+   * VoiceOver on Mac will read both
+   */
+  placeholder: PropTypes.string,
+
+  /**
+   * Rendered icon for the Search.
+   * Can be a React component class
+   */
+  renderIcon: PropTypes.oneOfType([PropTypes.func, PropTypes.object]),
+
+  /**
+   * Specify the role for the underlying `<input>`, defaults to `searchbox`
+   */
+  role: PropTypes.string,
+
+  /**
+   * Specify the size of the Search
+   */
+  size: PropTypes.oneOf(['sm', 'md', 'lg']),
+  /**
+   * Optional prop to specify the type of the `<input>`
+   */
+  type: PropTypes.string,
+
+  /**
+   * Specify the value of the `<input>`
+   */
+  value: PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
+};
+
+function CustomSearchIcon({ icon }) {
+  const prefix = usePrefix();
+
+  if (icon) {
+    return React.cloneElement(icon, {
+      className: `${prefix}--search-magnifier-icon`,
     });
-  };
-
-  handleKeyDown = (evt) => {
-    if (match(evt, keys.Escape)) {
-      this.clearInput(evt);
-    }
-  };
-
-  render() {
-    const {
-      className,
-      type,
-      id = (this._inputId =
-        this._inputId ||
-        `search__input__id_${Math.random().toString(36).substr(2)}`),
-      placeHolderText,
-      placeholder,
-      labelText,
-      closeButtonLabelText,
-      small,
-      size = !small ? 'xl' : 'sm',
-      light,
-      disabled,
-      onChange,
-      onKeyDown,
-      renderIcon,
-      onClear, // eslint-disable-line no-unused-vars
-      onExpand, // eslint-disable-line no-unused-vars, react/prop-types
-      ...other
-    } = this.props;
-
-    const { hasContent } = this.state;
-
-    const scope = this.context;
-    let enabled;
-
-    if (scope.enabled) {
-      enabled = scope.enabled('enable-v11-release');
-    }
-
-    return (
-      <PrefixContext.Consumer>
-        {(prefix) => {
-          const searchClasses = classNames({
-            [`${prefix}--search`]: true,
-            [`${prefix}--search--sm`]: size === 'sm',
-            // V11: change to md
-            [`${prefix}--search--lg`]: enabled ? size === 'md' : size === 'lg',
-            // V11: change to lg
-            [`${prefix}--search--xl`]: enabled ? size === 'lg' : size === 'xl',
-            [`${prefix}--search--light`]: light,
-            [`${prefix}--search--disabled`]: disabled,
-            [className]: className,
-          });
-
-          const clearClasses = classNames({
-            [`${prefix}--search-close`]: true,
-            [`${prefix}--search-close--hidden`]: !hasContent,
-          });
-
-          let customIcon;
-          if (renderIcon) {
-            customIcon = React.cloneElement(renderIcon, {
-              className: `${prefix}--search-magnifier-icon`,
-            });
-          }
-
-          const searchId = `${id}-search`;
-          const searchIcon = renderIcon ? (
-            customIcon
-          ) : (
-            <SearchIcon className={`${prefix}--search-magnifier-icon`} />
-          );
-
-          return (
-            <div
-              role="search"
-              aria-labelledby={searchId}
-              className={searchClasses}>
-              <div
-                className={`${prefix}--search-magnifier`}
-                ref={(magnifier) => {
-                  this.magnifier = magnifier;
-                }}>
-                {searchIcon}
-              </div>
-              <label id={searchId} htmlFor={id} className={`${prefix}--label`}>
-                {labelText}
-              </label>
-              <input
-                role="searchbox"
-                autoComplete="off"
-                {...other}
-                type={type}
-                disabled={disabled}
-                className={`${prefix}--search-input`}
-                id={id}
-                placeholder={placeHolderText || placeholder}
-                onChange={composeEventHandlers([onChange, this.handleChange])}
-                onKeyDown={composeEventHandlers([
-                  onKeyDown,
-                  this.handleKeyDown,
-                ])}
-                ref={(input) => {
-                  this.input = input;
-                }}
-              />
-              <button
-                className={clearClasses}
-                disabled={disabled}
-                onClick={this.clearInput}
-                type="button"
-                aria-label={closeButtonLabelText}>
-                <Close />
-              </button>
-            </div>
-          );
-        }}
-      </PrefixContext.Consumer>
-    );
   }
+  return <SearchIcon className={`${prefix}--search-magnifier-icon`} />;
 }
+
+CustomSearchIcon.propTypes = {
+  /**
+   * Rendered icon for the Search. Can be a React component class
+   */
+  icon: PropTypes.oneOfType([PropTypes.func, PropTypes.object]),
+};
+
+export default Search;
